Created CanvasPanel component
